Voilà, j'ai réussi les optimisations pour la raquette - j'ai optimiser le dessin de la raquette quand elle bouge - Maintenant la balle n'entre plus dans la raquette depuis les cotés

+++ workspace/ghostbuster_02/inc/racket.h
@@ -19,10 +19,10 @@
 
 // structure containing racket position, width and height
 typedef struct {
-	int x;
-	int y;
-	int width;
-	int height;
+	uint16_t x;
+	uint16_t y;
+	uint16_t width;
+	uint16_t height;
 } racket_t;
 
 // racket instance
